Fix tmp_data_path test

+++ query/src/configs/config_test.rs
@@ -76,7 +76,7 @@ rpc_tls_meta_service_domain_name = \"localhost\"
 storage_type = \"disk\"
 
 [storage.disk]
-data_path = \"\"
+data_path = \"_data\"
 temp_data_path = \"\"
 
 [storage.s3]
@@ -182,7 +182,11 @@ fn test_initial_dir() -> Result<()> {
     conf.initial_dir()?;
 
     // Remove.
-    std::fs::remove_dir_all(conf.storage.disk.data_path)?;
-    std::fs::remove_dir_all(conf.storage.disk.temp_data_path)?;
+    if !conf.storage.disk.data_path.is_empty() {
+        std::fs::remove_dir_all(conf.storage.disk.data_path)?;
+    }
+    if !conf.storage.disk.temp_data_path.is_empty() {
+        std::fs::remove_dir_all(conf.storage.disk.temp_data_path)?;
+    }
     Ok(())
 }
